+++ go-cosmwasm/src/lib.rs
@@ -260,7 +260,7 @@ pub extern "C" fn release_cache(cache: *mut cache_t) {
 
 #[repr(C)]
 pub struct EnclaveRuntimeConfig {
-    pub module_cache_size: u8,
+    pub module_cache_size: u32,
 }
 
 impl EnclaveRuntimeConfig {
